refactor: remove duplicated context
