+++ src/bot/commands/removerepo.ts
@@ -5,7 +5,7 @@ import { config } from "@/config";
 import { db, schema } from "@/db";
 import { isGitHubUrl } from "@/github";
 
-import type { BotContext } from "../context";
+import type { BotContext } from "../bot";
 
 import { escapeMarkdown } from "../../lib/escape-markdown";
